+++ NEWS.md
@@ -4,8 +4,8 @@
 
 * new function `plot_categorical()` (#47)
 * new colour palette `safe_pal()` and corresponding scales `scale_fill_safe()`
-  and `scale_colour_safe()` for `ggplot2`
-* new argument `minimal` for `plot_proportions()`
+  and `scale_colour_safe()` for `ggplot2` (#49)
+* new argument `minimal` for `plot_proportions()` (#51)
 
 **Minor breaking change**